+++ cmd/gaia/cmd/gaiad/main.go
@@ -2,8 +2,12 @@ package main
 
 import (
 	"encoding/json"
+	"io"
+
+	"github.com/cosmos/cosmos-sdk/baseapp"
 
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 
 	abci "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/libs/cli"
@@ -38,11 +42,13 @@ func main() {
 	}
 }
 
-func newApp(logger log.Logger, db dbm.DB) abci.Application {
-	return app.NewGaiaApp(logger, db)
+func newApp(logger log.Logger, db dbm.DB, traceStore io.Writer) abci.Application {
+	return app.NewGaiaApp(logger, db, traceStore, baseapp.SetPruning(viper.GetString("pruning")))
 }
 
-func exportAppStateAndTMValidators(logger log.Logger, db dbm.DB) (json.RawMessage, []tmtypes.GenesisValidator, error) {
-	gapp := app.NewGaiaApp(logger, db)
-	return gapp.ExportAppStateAndValidators()
+func exportAppStateAndTMValidators(
+	logger log.Logger, db dbm.DB, traceStore io.Writer,
+) (json.RawMessage, []tmtypes.GenesisValidator, error) {
+	gApp := app.NewGaiaApp(logger, db, traceStore)
+	return gApp.ExportAppStateAndValidators()
 }