+++ CHANGELOG.md
@@ -11,11 +11,14 @@ BREAKING CHANGES
 
 FEATURES
 * [baseapp] NewBaseApp now takes option functions as parameters
+* [store] Added support for tracing multi-store operations via `--trace-store`
+* [store] Pruning strategy configurable with pruning flag on gaiad start
 
 BUG FIXES
-* [server] 'gaiad init' command family now writes provided name as the moniker in `config.toml`
+* \#1630 - redelegation nolonger removes tokens from the delegator liquid account
 * [keys] \#1629 - updating password no longer asks for a new password when the first entered password was incorrect
 * [lcd] importing an account would create a random account
+* [server] 'gaiad init' command family now writes provided name as the moniker in `config.toml`
 
 ## 0.20.0
 
@@ -109,6 +112,7 @@ FEATURES
   - Auth has its invariants checked within the framework
 * [tests] Add WaitForNextNBlocksTM helper method
 * [keys] New keys now have 24 word recovery keys, for heightened security
+- [keys] Add a temporary method for exporting the private key
 
 IMPROVEMENTS
 * [x/bank] Now uses go-wire codec instead of 'encoding/json'